fix: next235 takes BIGINT, since a grid dimension can exceed 2^31
MAX_NF allows a fine-grid dimension up to 1e12, and set_nf_type12 hands one to
next235, whose long argument holds 32 bits on Windows. The wrapped value is
negative, the std::max(n, 1) inside clamps it to 1, and the call returns 2: a
silently tiny grid rather than a crash. good_size_235 already works in size_t,
so the narrow type only ever lived in the wrapper. fine_grid_len returns BIGINT
for the same reason.

testutils now rounds a dimension above 2^31, so the guard runs on the platform
that has the bug.

+++ src/common/utils.cpp
@@ -115,10 +115,10 @@ static size_t good_size_235(size_t n, size_t required_factor) {
   return good_size_235((n + required_factor - 1) / required_factor) * required_factor;
 }
 
-long next235(long n, long required_factor) {
-  n               = std::max<long>(n, 1);
-  required_factor = std::max<long>(required_factor, 1);
-  return good_size_235(size_t(n), size_t(required_factor));
+BIGINT next235(BIGINT n, BIGINT required_factor) {
+  n = std::max<BIGINT>(n, 1);
+  required_factor = std::max<BIGINT>(required_factor, 1);
+  return BIGINT(good_size_235(size_t(n), size_t(required_factor)));
 }
 
 } // namespace common
